Extract CollectionItemActions component and refine item layout
