+++ app/modules/featureFlags/helpers/getAddUsersToFeatureFlagDialogItemActions.tsx
@@ -1,4 +1,4 @@
-import type { CollectionItemAction } from "@/components/ui/collectionItemContent";
+import type { CollectionItemAction } from "@/components/ui/collectionItemActions";
 import type { User } from "~/modules/users/users.types";
 
 export default function getAddUsersToFeatureFlagDialogItemActions(
