+++ app/modules/teams/helpers/getTeamPromptsItemActions.tsx
@@ -1,4 +1,4 @@
-import type { CollectionItemAction } from "@/components/ui/collectionItemContent";
+import type { CollectionItemAction } from "@/components/ui/collectionItemActions";
 
 export default function getTeamPromptsItemActions(): CollectionItemAction[] {
   return [];
