+++ app/uikit/components/ui/collectionItemContent.tsx
@@ -1,23 +1,8 @@
 import map from "lodash/map";
-import { EllipsisVertical } from "lucide-react";
-import React, { type ReactElement } from "react";
+import { type ReactElement } from "react";
 import { Badge } from "./badge";
-import { Button } from "./button";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "./dropdown-menu";
-import { ItemActions, ItemContent, ItemDescription, ItemTitle } from "./item";
-
-export type CollectionItemAction = {
-  icon?: ReactElement;
-  action: string;
-  text: string;
-  variant?: "default" | "destructive" | undefined;
-};
+import type { CollectionItemAction } from "./collectionItemActions";
+import { ItemContent, ItemDescription, ItemTitle } from "./item";
 
 export type CollectionItemMeta = {
   text: string;
@@ -33,7 +18,7 @@ export type CollectionItemAttributes = {
   isDisabled?: boolean;
 };
 
-type CollectionItemProps = {
+export type CollectionItemProps = {
   id: string;
   title: string;
   description?: string;
@@ -50,108 +35,32 @@ type CollectionItemProps = {
   }) => void;
 };
 
-const CollectionItemActions = ({
-  id,
-  actions,
-  isDisabled,
-  onItemActionClicked,
-}: Pick<
-  CollectionItemProps,
-  "id" | "actions" | "isDisabled" | "onItemActionClicked"
->) => {
-  if (actions.length === 0) return null;
-  return (
-    <ItemActions>
-      {(actions.length > 1 && (
-        <DropdownMenu>
-          <DropdownMenuTrigger asChild disabled={isDisabled}>
-            <Button
-              variant="ghost"
-              className="data-[state=open]:bg-muted text-muted-foreground flex size-8"
-              size="icon"
-            >
-              <EllipsisVertical />
-              <span className="sr-only">Open menu</span>
-            </Button>
-          </DropdownMenuTrigger>
-          <DropdownMenuContent align="end" className="w-fit">
-            {map(actions, (action, index) => {
-              return (
-                <React.Fragment key={action.action}>
-                  <DropdownMenuItem
-                    variant={action.variant}
-                    onClick={(event) => {
-                      event.stopPropagation();
-                      event.preventDefault();
-                      if (onItemActionClicked) {
-                        onItemActionClicked({ id, action: action.action });
-                      }
-                    }}
-                  >
-                    {action.icon ? action.icon : null}
-                    {action.text}
-                  </DropdownMenuItem>
-                  {index !== actions.length - 1 && <DropdownMenuSeparator />}
-                </React.Fragment>
-              );
-            })}
-          </DropdownMenuContent>
-        </DropdownMenu>
-      )) || (
-        <Button
-          variant="ghost"
-          onClick={(event) => {
-            event.stopPropagation();
-            event.preventDefault();
-            if (onItemActionClicked) {
-              onItemActionClicked({ id, action: actions[0].action });
-            }
-          }}
-        >
-          {actions[0].icon ? actions[0].icon : null}
-          {actions[0].text}
-        </Button>
-      )}
-    </ItemActions>
-  );
-};
-
 const CollectionItemContent = ({
-  id,
   title,
   description,
   meta = [],
-  actions = [],
-  isDisabled = false,
-  onItemActionClicked,
-}: Omit<CollectionItemProps, "to">) => (
-  <>
-    <ItemContent className="gap-1">
-      <ItemTitle className="text-lg">{title}</ItemTitle>
-      <ItemDescription>{description}</ItemDescription>
-      <div className="flex w-full flex-wrap gap-2">
-        {map(meta, (metaItem, index) => {
-          return (
-            <Badge
-              key={index}
-              variant="outline"
-              className="text-muted-foreground"
-            >
-              {metaItem.icon && metaItem.icon}
-              {metaItem.text}
-            </Badge>
-          );
-        })}
-      </div>
-    </ItemContent>
-    <CollectionItemActions
-      id={id}
-      actions={actions}
-      isDisabled={isDisabled}
-      onItemActionClicked={onItemActionClicked}
-    />
-  </>
+}: Omit<
+  CollectionItemProps,
+  "to" | "id" | "actions" | "isDisabled" | "onItemActionClicked"
+>) => (
+  <ItemContent className="gap-1">
+    <ItemTitle className="text-lg">{title}</ItemTitle>
+    <ItemDescription>{description}</ItemDescription>
+    <div className="flex w-full flex-wrap gap-2">
+      {map(meta, (metaItem, index) => {
+        return (
+          <Badge
+            key={index}
+            variant="outline"
+            className="text-muted-foreground"
+          >
+            {metaItem.icon && metaItem.icon}
+            {metaItem.text}
+          </Badge>
+        );
+      })}
+    </div>
+  </ItemContent>
 );
 
-export { CollectionItemActions };
 export default CollectionItemContent;
